+++ ICD.Connect.Panels.CrestronPro/TriListAdapters/Xpanel/XPanelForSmartGraphicsAdapter.cs
@@ -1,7 +1,6 @@
 ﻿#if SIMPLSHARP
 using Crestron.SimplSharpPro;
 using Crestron.SimplSharpPro.UI;
-
 #endif
 
 namespace ICD.Connect.Panels.CrestronPro.TriListAdapters.Xpanel
@@ -13,6 +12,18 @@ namespace ICD.Connect.Panels.CrestronPro.TriListAdapters.Xpanel
 #endif
 	{
 #if SIMPLSHARP
+		/// <summary>
+		/// Gets the current online status of the panel.
+		/// </summary>
+		/// <returns></returns>
+		protected override bool GetIsOnlineStatus()
+		{
+			// Always show XPanels as online (ignore Crestron IsOnline feedback) because
+			// they're used for diagnostics and we don't want to pollute telemetry every
+			// time someone connects/disconnects
+			return Panel != null;
+		}
+
 		/// <summary>
 		/// Creates an instance of the wrapped trilist.
 		/// </summary>